Refactor imports and simplify module exports
