+++ src/ai-function/ai-extract-function.ts
@@ -1,6 +1,6 @@
 import { type z } from 'zod';
 
-import { type Model, MsgUtil } from '../model/index.js';
+import { type Model, MsgUtil } from '../index.js';
 import { createAIFunction } from './ai-function.js';
 import { createAIRunner } from './ai-runner.js';
 import { type ExtractFunction } from './types.js';
