--- src/ai-function/index.ts
@@ -1,8 +1,0 @@
-export { createAIExtractFunction } from './ai-extract-function.js';
-export { createAIFunction } from './ai-function.js';
-export { createAIRunner, handleFunctionCallMessage } from './ai-runner.js';
-export {
-  type AIFunction,
-  type AIRunner,
-  type ExtractFunction,
-} from './types.js';
