+++ src/index.ts
@@ -1,3 +1,32 @@
-// This is intentionally empty, but here to make bundlers happy.
-
-export {};
+export { createAIExtractFunction } from './ai-function/ai-extract-function.js';
+export { createAIFunction } from './ai-function/ai-function.js';
+export {
+  createAIRunner,
+  handleFunctionCallMessage,
+} from './ai-function/ai-runner.js';
+export {
+  type AIFunction,
+  type AIRunner,
+  type ExtractFunction,
+} from './ai-function/types.js';
+export { createExtractFunction } from './extract/index.js';
+export type { ChatModelArgs } from './model/chat.js';
+export { ChatModel } from './model/chat.js';
+export { createOpenAIClient } from './model/clients/openai.js';
+export type { CompletionModelArgs } from './model/completion.js';
+export { CompletionModel } from './model/completion.js';
+export type { EmbeddingModelArgs } from './model/embedding.js';
+export { EmbeddingModel } from './model/embedding.js';
+export type { ModelArgs } from './model/model.js';
+export { AbstractModel } from './model/model.js';
+export type { SparseVectorModelArgs } from './model/sparse-vector.js';
+export { SparseVectorModel } from './model/sparse-vector.js';
+export type { Model, Msg } from './model/types.js';
+export { type CacheKey, type CacheStorage } from './model/utils/cache.js';
+export { calculateCost } from './model/utils/calculate-cost.js';
+export { AbortError, RefusalError } from './model/utils/errors.js';
+export { extractJsonObject } from './model/utils/extract-json.js';
+export { extractZodObject } from './model/utils/extract-zod-object.js';
+export { MsgUtil } from './model/utils/message-util.js';
+export { createTokenizer, type Tokenizer } from './model/utils/tokenizer.js';
+export { zodToJsonSchema } from './model/utils/zod-to-json-schema.js';
