+++ src/swarm/swarm-tools.ts
@@ -5,7 +5,7 @@ import {
   type Msg,
   MsgUtil,
   zodToJsonSchema,
-} from '../model/index.js';
+} from '../index.js';
 import { getErrorMsg } from '../model/utils/errors.js';
 import { cleanString } from '../model/utils/message-util.js';
 import { type Agent, type SwarmFunc } from './types.js';
